Only call taskkill if windows fix path for simulator if myro not installed, but in cwd
git-svn-id: http://svn.cs.brynmawr.edu/Myro@233 ed25d823-a23d-414a-8cb8-cff62a94a292

+++ trunk/Start Python.pyw
@@ -2,7 +2,8 @@
 
 from idlelib import PyShell
 import sys, os
-os.system("taskkill /IM pythonw.exe")
+if os.name in ['nt', 'dos', 'os2'] :
+    os.system("taskkill /IM pythonw.exe")
 #sys.argv = [sys.argv[0]] + ['-n'] 
 try:
     commands = open("mystart.py", "r").readlines()
